Fixes casting issues

+++ PGSuperAppPlugin/BridgeDescFramingPage.cpp
@@ -300,7 +300,7 @@ void CBridgeDescFramingPage::OnLayoutBySpanLengths()
 
    CSpanLengthDlg dlg;
    dlg.m_SpanLengths = spanLengths;
-   dlg.m_PierIdx = m_StartingPierNumber-1;
+   dlg.m_PierIdx = (int)(m_StartingPierNumber-1);
    if ( dlg.DoModal() == IDOK )
    {
       m_Grid.SetSpanLengths(dlg.m_SpanLengths,dlg.m_PierIdx - m_StartingPierNumber + 1);
